Revert "Configuration for Touchpoint services - Salesforce and Zuora."
