+++ app/Global.scala
@@ -2,11 +2,11 @@ import filters.CheckCacheHeadersFilter
 import monitoring.SentryLogging
 import play.api.Application
 import play.api.mvc.WithFilters
-import services.TouchpointBackend
+import services.SalesforceRepo
 
 object Global extends WithFilters(CheckCacheHeadersFilter) {
   override def onStart(app: Application) {
     SentryLogging.init()
-    TouchpointBackend.All.foreach(_.start())
+    SalesforceRepo.salesforce.authTask.start()
   }
 }
