+++ conf/DEV.conf
@@ -14,6 +14,4 @@ subscriptions.url="https://sub.thegulocal.com"
 
 play.ws.acceptAnyCertificate=true
 
-touchpoint.backend.default=DEV
-
-sentry.dsn = ""
+sentry.dns = ""
