+++ conf/PROD.conf
@@ -10,5 +10,3 @@ identity {
 }
 
 subscriptions.url="https://subscribe.theguardian.com"
-
-touchpoint.backend.default=PROD
