+++ conf/touchpoint.DEV.conf
@@ -13,12 +13,5 @@ touchpoint.backend.environments {
         token=""
       }
     }
-    zuora {
-      api {
-        url = "https://apisandbox.zuora.com/apps/services/a/58.0"
-        username = ""
-        password = ""
-      }
-    }
   }
 }
